refactor: rename `trace` module to `telemetry`
